Adding slow typing, slowly
Something is wrong in main and here. Trying to hunt down, may need to revert changes done recently.

Slow typing of text is being attempted to add here, but the program doesn't seem to run 9/10 times, then does sometimes. Having to make the logging system more robust to assist in this.